Still moving things around and improving tomcat8_* scripts.

+++ tomcat8/ubuntu/tomcat8_uninstall.sh
@@ -1,5 +1,10 @@
 #!/usr/bin/env bash
 
+if [[ $EUID -ne 0 ]]; then
+   echo "This script must be run as root"
+   exit 1
+fi
+
 service tomcat8 stop
 
 rm -rf /usr/share/tomcat8
